+++ CHANGES.rst
@@ -4,9 +4,10 @@ Changelog
 0.1.4 (unreleased)
 ------------------
 
-- Issue#3 `Fixed  typo <https://github.com/nazrulworld/collective.recipe.vscode/issues/3>`_
-- Issue#4 `Documented isort <https://github.com/nazrulworld/collective.recipe.vscode/issues/4>`_
-- Issue#5 `Automatic .env file generation`_
+- Issue#3 `Fixed  typo <https://github.com/nazrulworld/collective.recipe.vscode/issues/3>`_ [parruc]
+- Issue#4 `Documented isort <https://github.com/nazrulworld/collective.recipe.vscode/issues/4>`_ [parruc]
+- Issue#5 `Automatic .env file generation <https://github.com/nazrulworld/collective.recipe.vscode/issues/5>`_ [parruc]
+
 
 
 0.1.3 (2019-03-12)